multithread bug fix: bbl fetch logic

+++ src/tracer/sinuca/utils/dynamic_trace_reader.hpp
@@ -73,7 +73,7 @@ class DynamicTraceReader {
     inline int GetRecordType() {
         return this->recordArray[this->recordArrayIndex].recordType;
     }
-    inline unsigned long GetBasicBlockIdentifier() {
+    inline unsigned int GetBasicBlockIdentifier() {
         return this->recordArray[this->recordArrayIndex]
             .data.basicBlockIdentifier;
     }
